backend/requirements.txt
# Django core
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0

# WebSockets
channels>=4.1,<5.0
daphne>=4.1,<5.0

# Database
psycopg2-binary>=2.9,<3.0
pgvector>=0.3,<1.0

# LLM
litellm>=1.40,<2.0

# Math / Solver
numpy>=1.26,<3.0
scipy>=1.14,<2.0
pulp>=2.8,<3.0

# RAG / Document processing
pdfplumber>=0.11,<1.0
tiktoken>=0.7,<1.0

# Dev tools
ruff>=0.5,<1.0
black>=24.0,<25.0
isort>=5.13,<6.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
